pyproject.toml
[project]
name = "restjango"
version = "0.1.0"
description = ""
authors = [
    {name = "Raz",email = "razibh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]

[tool.poetry]
package-mode = false
packages = [{include = "restjango", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "django-stubs[compatible-mypy] (>=5.2.8,<6.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "djangorestframework-stubs[compatible-mypy] (>=3.16.6,<4.0.0)"
]

[tool.ruff]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "restjango.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "restjango.settings"